Added error handling to tex cmd execution

+++ lib/latex_to_pdf.rb
@@ -1,32 +1,40 @@
 # encoding: utf-8
 class LatexToPdf
   @@options = { tex_path: '/usr/bin', tex_engine: 'pdflatex', parse_twice: false }
-  class << self
 
+  class << self
     # Converts a string of LaTeX +code+ into a binary string of PDF.
     #
     # pdflatex is used to convert the file and creates the directory +#{Rails.root}/tmp/rails-latex+ to store intermediate
     # files.
     def generate_pdf(code, options = {})
-      options.reverse_merge @@options
+      options.reverse_merge! @@options
 
       dir   = File.join(Rails.root, 'tmp', 'rails-latex', "#{Process.pid}-#{Thread.current.hash}")
       input = File.join(dir, 'input.tex')
       FileUtils.mkdir_p(dir)
 
       File.open(input, 'wb') {|io| io.write(code) }
-      tex_command = File.join(options[:tex_path], options[:tex_engine])
+      tex_command = File.join( options[:tex_path], options[:tex_engine] )
+      raise "LaTeX command '#{tex_command}' not found" unless File.exist?(tex_command)
+
       (options[:parse_twice] ? 2 : 1).times {
-        `#{tex_command} -output-directory #{dir} -interaction batchmode #{input}`
+        output = `PATH=#{options[:tex_path]} #{tex_command} -output-directory #{dir} -interaction nonstopmode #{input} 2>&1`
+        raise "#{options[:tex_command]} error: #{output}" if output.include?("Error")
         # :umask => 7, :out => :close, :err => :close, :in => :close) # not supported in ruby 1.8
       }
-      FileUtils.mv( input.sub(/\.tex$/, '.log'), File.join(dir, '..', 'input.log') )
+
+      if File.exist? log_file = input.sub(/\.tex$/, '.log')
+        FileUtils.mv(log_file, File.join(dir, '..', 'input.log') )
+      else
+        raise "#{options[:tex_command]} failed: Log file #{log_file} not found."
+      end
 
       if File.exist?( pdf_file = input.sub(/\.tex$/, '.pdf') )
         result = File.read(pdf_file)
         FileUtils.rm_rf(dir)
       else
-        raise "pdflatex failed: See #{input.sub(/\.tex$/, '.log')} for details"
+        raise "#{options[:tex_command]} failed: See #{input.sub(/\.tex$/, '.log')} for details"
       end
       result
     end
@@ -70,8 +78,12 @@ class LatexToPdf
       @latex_escaper.latex_esc(text.to_s).html_safe
     end
 
-    def settings(options_hash)
-      @@options.merge options_hash
+    def settings
+      @@options
+    end
+
+    def settings=(options_hash)
+      @@options.merge! options_hash
     end
   end
 end
